The atom storage field can be a short.

+++ gdk/gdk.h
@@ -1873,7 +1873,7 @@ gdk_export BAT *BBPquickdesc(bat b, int delaccess);
 typedef struct {
 	/* simple attributes */
 	char name[IDLENGTH];
-	int storage;		/* stored as another type? */
+	short storage;		/* stored as another type? */
 	short linear;		/* atom can be ordered linearly */
 	short size;		/* fixed size of atom */
 	short align;		/* alignment condition for values */
